fix: api 타입 오류 수정
